Remove native scan code from CKeyboardKey Remove isPressed status from CKeyboardKey Switched key value to Qt::Key
There is no way on OSX to get the native scan code of a key.
Therefore it is removed from this class, because we might
create dependent code which will not run on OSX.
CKeyboardKey is a abstraction of platform keys. It was representing
also the status when CKeyboardKey was sent in signals. So this
can be removed. Pressed status is sent as argument to the registered
method.

refs #83