Fix launching of shipped dbus-daemon
When using QProcess::startDetached on Windows, a console Window is
visible. Since dbus-daemon should run silently in the background,
use our customized CProcess instead which supports starting without
console.
This commit also fixes a race condition. CDBusServer launched the
dbus-daemon and immediatly tried to register. If dbus-daemon was not yet
initialized and ready to accept connections, registering the service
failed. This is fixed by a small timeout.

refs #615