Conditionally limit UI login modes in case no simulator is attached
Currently the logic to fall back to observer login, in case no
simulator is attached, is implemented in core. Apart from a log message
there was no clear feedback to the user what has happened.
With this commit, the pilot related login modes are disabled if no
simulator is attached. A tool tip explains the reason.
This is enabled only for shipped versions.

refs #872