Make installing of ICU libraries optional on Linux
ICU libraries are shipped in the official Qt installers but in a
local or distribution build, Qt links against system libraries which made
the install target fail in this two cases.
By checking whether ICU libraries are inside the Qt library dir,
we make sure that we redistribute those non system ICU libraries.