Remove deprecated custom packet API from INetwork
Usage of this API was already discouraged, since all official
VATSIM packet handling should be in vatlib itself or for custom defined
packets in CNetworkVatlib. If a new custom packet needs to be added,
it should go into CNetworkVatlib directly and made available via normal
API methods to abstract the packets custom nature. The fact that
custom packets exist, should be kept internal. However, a debug message
is raised in case of an unknown custom packet in order to allow
detection and debugging.

refs #870