Pass application list directly to CHotkeyDialog constructor
The list of applications is already known when the CHotkeyDialog
object is constructed. So there was no good reason to defer this
to a later time with a setter.
In contrast, supporting a setter function was much more complex, since
internally all GUI signals were already connected and more carefull
case handling would be necessary (e.g. the machine combobox would
change its index). Therefore the setter is removed in favor of passing
the application list directly to the constructor.
This fixes also an issue that the machine name was not properly set up
in the dialog when editing a hotkey.
This also reverts the change from CIdentifier to string based machine
handling.

refs #881