Refactor CProcessCtrl to start detached with or without console
QTBUG-53833 - which caused a console application to launch with a console
if it was started with QProcess::startDetached - is fixed in 5.8.0.
There were use cases for both options - with and without console -
but not offered by Qt API. Therefore CProcessCtrl is refactored  as a
wrapper to offer both options. If no console is required, it defaults to
QProcess::startDetached. Otherwise uses its own implementation.

refs #866