Split listening for simulators and loading their plugins
Previously loadSimulatorPlugin was doing two things. First starting
a listener and second loading the plugin itself. This was confusing
since the sequence is to listen first and load not before a simulator
is available.
This change splits this up properly into two methods. listenForSimulator()
is called when the plugin is started and as soon as a simulator is started,
its plugin will be loaded.

refs #909