Add first batch of configuration tests
Summary:
Configuration tests check whether all mandatory dependencies to build
swift are met. They can also be used to check optional dependencies, which
results in certain features to be enabled or disabled.
This first batch tests for mandatory libraries libpng, zlib and OpenGL.
If this is not tested and some libraries are missing, the developer will
be notified very late in the build.

Reviewers: #swift_pilot_client, msutcliffe

Reviewed By: #swift_pilot_client, msutcliffe

Subscribers: msutcliffe, jenkins

Differential Revision: https://dev.swift-project.org/D34