Generate DBus session.conf from template for each platform individually
Before this commit we used to have a static session.conf. But the session
configuration needs to be different for Mac OS and Windows. Our static
one was Windows only. This caused the packaged DBus daemon to fail on
Mac OS.
This commit solves this by generating the config file at qmake time with
special configuration for each operating system.

ref T170