Remove the test connection and custom launch of dbus daemon
We cannot easily launch the dbus session ourselves since all other
applications don't know how to connect to it. We need to rely on the
supported solutions on different platforms. If the session bus is not
available, the workarounds would be huge to manually launch it and tell
all clients the session address.

ref T170