Fix building a bitrock installer with MinGW binaries
Summary:
This simply fixes the errors, one will see if an installer is created after
a MinGW build. This does not change the fact that a MinGW installer is
not supported and not necessary. The installed software will not be able
to run due to missing dependencies. But the installer can be created and
executed for testing.

Reviewers: #swift_pilot_client, kbasan

Reviewed By: #swift_pilot_client, kbasan

Subscribers: jenkins

Differential Revision: https://dev.swift-project.org/D67