Limit the number of parallel dot threads
If doxygen is executing dot to produce graphics, it will run it with as
many threads as physical cores are available. On Jenkins, when running
inside a LXC, the number of physical cores is much higher than the
ones assigned to the container. For increased the load inside the container
during the dot run inefficiently. Therefore limit the number of threads
to 2.