[GFS] Properly validate cloud pressure levels
For some reason, the parser returned 0.0 instead of 9.998e20 for undefined values.
Atmospheric pressure levels cannot be lower than 1000, hence use this as the lower boundary.
In all other circumstances, the pressure level needs to be set (either to real value or to nan) for the follow up code to detect undefined values.