Improved the readability of the loadCslPackages function.
Summary:
In this function, a local class `Prefix` is used to represent the path
of a CSL package and to test whether a model path is a subpath of that
path. To avoid false positives, a trailing `/` character is appended.

This lead to confusing code where a `/` is appended in the first loop
and then chopped in the second loop. Instead, encapsulate the append
and the chop inside the `Prefix` class.

Differential Revision: https://dev.swift-project.org/D107