[xswiftbus] Fix UTF-8 bug when displaying text messages
std::string is not UTF-8 aware, so it was possible to create
malformed strings by splitting in the middle of a code point.

The splitting code also did read beyond the end of the input string.

Now we use a Unicode-aware iterator adaptor to help find the
correct place to split the string across multiple lines.