[xswiftbus] Workaround for shortcoming of qmake INSTALLS wildcard expansion
When using wildcards in the `files` member of an `INSTALLS` variable,
the files must exist when qmake is run or the wildcards won't be expanded.
The libevent and libpng libraries are copied to `$$DestRoot/lib` as a
post-build step in `blackmisc.pro`, so I work around with a shell command.