refactor: Remove navigation equipment from remark generator
The NAV/ field in the remarks (same as COM/) is a freetext field. The content
may be dependent on the air traffic authority/country.
Further, most of the equipment information is nowadays handled with the
ICAO equipment code. NAV/ is likely required in special cases only, for
example when filing equipment code Z (Other equipment).