refactor: Remove CUrlList
Nowadays most of the loadbalancing is done on the server-side and hence
there is only a single datafile URL (and other URLs) inside the
boostrap.json.
The features of the CUrlList are hence not really used. This is also a
step into removing CUrl and using QUrl instead, to avoid maintaining a
separate URL class.