fix(qss): Set menu background color
As a black text color is enforced for the menu, we should also enforce a
white background for a good contrast.
This is more or less a workaround as this does not respect any system
style settings (dark mode, ...).
This fix will become obsolete once tackling #138 (removing/reducing stylesheet).

Fixes #139