refactor: Make protected IContext members private
According to C++ Core Guideline C.133, protected data should be avoided
because they introducing complexity and bugs. In this case for example,
the implementing contexts could just change the m_debugEnabled member
instead calling setDebugEnabled() if not paying attention.
setDebugEnabled() also emit a signal on change, which is accidentially
not emitted otherwise.