refactor: Remove "connected" LED from COM panel
The LED is a leftover from the pre-AFV voicerooms. The LED does not
light up when connected to UNICOM or CTAF frequencies which still can be
used to transmit & receive voice. This also removes the functionality
that the currently tuned station was displayed when hovering over the
"COMx" label. This feature was not documented and not known by many
users.
By removing these features, one context method is no longer required,
which further decouples the network context from the OwnAircraft context.