refactor: Remove COM sync button from cockpit page
This setting is simulator specific and can already be adjusted from the
simulator plugin settings. To avoid complexity, it should only be
adjustable from one location. As this setting doesn't need to be
adjusted often inflight, it should be fine to not have it directly
accessible.
This also fixes a previous issue, where a change of the setting within
the simulator plugin settings did not propagated to the cockpit page.