refactor: Remove log messages from status bar
With the new warn/error button, the user already is informed if
something important is logged. By only showing nearly all log messages
in the status bar, messages can be missed easily.
Fixes also an issue where the status bar was accessed before it was
initialized (crash with assert). (Fixes #299).