refactor: Remove manual high-dpi support
This removes the --scale command line option.
Qt6 seems to have this already integrated to scale according to the
nativ display settings: https://doc.qt.io/qt-6.8/highdpi.html
Adjusting the scale on a per-application basis is not intended (anymore)
according to the documentation.
Tested on a (near) high-dpi screen.